requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
airstack==0.0.7
async-timeout==4.0.3
attrs==23.2.0
charset-normalizer==3.3.2
frozenlist==1.4.1
graphql-core==2.3.1
idna==3.6
multidict==6.0.5
numpy==1.26.4
pandas==2.2.1
promise==2.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
Rx==1.6.3
six==1.16.0
tzdata==2024.1
yarl==1.9.4
